+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG.md
 
 ## 2.1.1 - 2022-05-13
+### Features (1 change)
+- Top elements in product and content search result
+
 ### Fix (1 change)
 
 - Categories with product stream are not longer loaded via ff
+- Category names can now have spaces
 
 ## Template